Allow to set filter on interactive task list
